Move ObtainFastaRegions into pipeline_utils and setup v2
